♻️ [Refactor] TodoCategryEditView 구조 개선 #186
- CategoryEditCell 분리
- TodoCategryEditView 구조 개선
