+++ Indayvidual/Indayvidual/Sources/Features/TodoList/View/CategoryEditCell.swift
@@ -1,0 +1,64 @@
+//
+//  CategoryEditCell.swift
+//  Indayvidual
+//
+//  Created by 김지민 on 8/20/25.
+//
+
+import SwiftUI
+
+// MARK: - CategoryEditCell
+struct CategoryEditCell: View {
+    let category: Category
+    let isPendingDelete: Bool
+    let onEdit: () -> Void
+    let onDelete: () -> Void
+
+    var body: some View {
+        HStack(spacing: 0) {
+            ZStack {
+                RoundedRectangle(cornerRadius: 16)
+                    .fill(Color.white)
+                    .frame(height: 48)
+
+                HStack(spacing: 12) {
+                    HStack(spacing: 7) {
+                        Image("categoryIcon")
+                            .resizable()
+                            .frame(width: 17, height: 17)
+                        Text(category.name)
+                            .font(.pretendSemiBold15)
+                            .foregroundColor(.gray900)
+                            .lineLimit(1)
+                            .minimumScaleFactor(0.8)
+                            .strikethrough(isPendingDelete, color: .red)
+                    }
+                    .padding(.horizontal, 14)
+                    .padding(.vertical, 6)
+                    .background(
+                        RoundedRectangle(cornerRadius: 10)
+                            .fill(category.color)
+                    )
+
+                    Spacer()
+
+                    Menu {
+                        Button("수정하기", action: onEdit)
+                        Button("삭제하기", role: .destructive, action: onDelete)
+                    } label: {
+                        Image(systemName: "ellipsis")
+                            .rotationEffect(.degrees(90))
+                            .foregroundColor(.gray)
+                            .frame(width: 32, height: 32)
+                    }
+                    .background(Color.white)
+                    .cornerRadius(8)
+                    .padding(.trailing, 8)
+                }
+                .padding(.leading, 14)
+                .padding(.trailing, 6)
+            }
+        }
+        .padding(.horizontal, 20)
+    }
+}
